608-Capturing-Logs-From-Containers-Running-On-ECS/cdk-AWS-Cookbook-408/requirements.txt
-e .
boto3
botocore
jsii
attr
attrs
cattrs
constructs
typing_extensions
jmespath
publication
python-dateutil
s3transfer
six
typing-extensions
urllib3
aws_cdk.core
aws-cdk.assets
aws-cdk.aws-apigateway
aws-cdk.aws-apigatewayv2
aws-cdk.aws-applicationautoscaling
aws-cdk.aws-autoscaling
aws-cdk.aws-autoscaling-common
aws-cdk.aws-autoscaling-hooktargets
aws-cdk.aws-certificatemanager
aws-cdk.aws-cloudformation
aws-cdk.aws-cloudfront
aws-cdk.aws-cloudwatch
aws-cdk.aws-codeguruprofiler
aws-cdk.aws-cognito
aws-cdk.aws-ecr
aws-cdk.aws-efs
aws-cdk.aws-elasticloadbalancing
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-events
aws-cdk.aws-iam
aws-cdk.aws-kms
aws-cdk.aws-lambda
aws-cdk.aws-logs
aws-cdk.aws-route53
aws-cdk.aws-route53-targets
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws-cdk.aws-sam
aws-cdk.aws-secretsmanager
aws-cdk.aws-servicediscovery
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
aws-cdk.aws-sqs
aws-cdk.aws-ssm
aws-cdk.cloud-assembly-schema
aws-cdk.custom-resources
aws-cdk.cx-api
aws-cdk.region-info
aws_cdk.aws_ec2
aws_cdk.aws_ecs
